chore: VideoEditLightFragment 삭제 후 VideoEditFragment로 대체
