--- app/src/main/res/layout/fragment_video_edit_light.xml
@@ -1,68 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<layout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools">
-
-    <data>
-
-    </data>
-
-    <androidx.constraintlayout.widget.ConstraintLayout
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        tools:context=".ui.videoedit.VideoEditFragment">
-
-
-        <com.google.android.exoplayer2.ui.AspectRatioFrameLayout
-            android:id="@+id/frame_video"
-            android:layout_width="match_parent"
-            android:layout_height="400dp"
-            android:padding="8dp"
-            app:layout_constraintTop_toTopOf="parent"
-            app:resize_mode="fixed_height">
-
-            <com.google.android.exoplayer2.ui.PlayerView
-                android:id="@+id/exoplayer"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:layout_gravity="center"
-                app:resize_mode="fixed_height" />
-        </com.google.android.exoplayer2.ui.AspectRatioFrameLayout>
-
-        <com.alphamovie.lib.AlphaMovieView
-            android:id="@+id/alphaView"
-            android:layout_width="match_parent"
-            android:layout_height="0dp"
-            app:accuracy="0.05"
-            app:alphaColor="#000000"
-            app:layout_constraintBottom_toBottomOf="@id/frame_video"
-            app:layout_constraintTop_toTopOf="@id/frame_video" />
-
-        <EditText
-            android:hint="제목을 입력해주세요"
-            android:id="@+id/et_memo_title"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:inputType="text"
-            app:layout_constraintTop_toBottomOf="@id/frame_video" />
-
-        <TextView
-            android:id="@+id/tv_subvideos"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_margin="16dp"
-            android:textSize="16sp"
-            app:layout_constraintTop_toBottomOf="@id/et_memo_title" />
-
-        <Button
-            android:id="@+id/btn_save_memo"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_margin="16dp"
-            android:text="저장하기"
-            app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintEnd_toEndOf="parent" />
-
-
-    </androidx.constraintlayout.widget.ConstraintLayout>
-</layout>
