+++ app/src/main/res/navigation/navigation.xml
@@ -71,7 +71,7 @@
             app:argType="string" />
         <action
             android:id="@+id/action_videoDoodleFragment_to_videoEditLightFragment"
-            app:destination="@id/videoEditLightFragment">
+            app:destination="@id/videoEditFragment">
             <argument
                 android:name="baseVideo"
                 app:argType="string" />
@@ -83,17 +83,12 @@
     </fragment>
 
     <fragment
-        android:id="@+id/videoEditFragment"
-        android:name="com.kotlinisgood.boomerang.ui.videoedit.VideoEditFragment"
-        android:label="fragment_video_edit"
-        tools:layout="@layout/fragment_video_edit" />
-    <fragment
         android:id="@+id/videoDoodleLightFragment"
         android:name="com.kotlinisgood.boomerang.ui.videodoodlelight.VideoDoodleLightFragment"
         android:label="VideoDoodleLightFragment" >
         <action
             android:id="@+id/action_videoDoodleLightFragment_to_videoEditLightFragment"
-            app:destination="@id/videoEditLightFragment">
+            app:destination="@id/videoEditFragment">
             <argument
                 android:name="baseVideo"
                 app:argType="string" />
@@ -106,8 +101,8 @@
             app:argType="string" />
     </fragment>
     <fragment
-        android:id="@+id/videoEditLightFragment"
-        android:name="com.kotlinisgood.boomerang.ui.videoeditlight.VideoEditLightFragment"
+        android:id="@+id/videoEditFragment"
+        android:name="com.kotlinisgood.boomerang.ui.videoedit.VideoEditFragment"
         android:label="VideoEditLightFragment" >
         <argument
             android:name="subVideos"
